requirements.txt
streamlit
groq
pandas
python-docx